commit avant ajout de repas daté

+++ src/main/java/fr/treeptik/controller/ExerciceController.java
@@ -53,7 +53,7 @@ public class ExerciceController {
 	}
 
 	@RequestMapping(method = RequestMethod.GET, value = "/update/{id}")
-	public String update(Model model, @PathVariable("id") Integer id)
+	public String update(Model model, @PathVariable("id") Long id)
 			throws ControllerException {
 		logger.info("--update ExerciceController-- exerciceId : " + id);
 
@@ -70,7 +70,7 @@ public class ExerciceController {
 	}
 
 	@RequestMapping(method = RequestMethod.GET, value = "/delete/{id}")
-	public String delete(Model model, @PathVariable("id") Integer id)
+	public String delete(Model model, @PathVariable("id") Long id)
 			throws ControllerException {
 		logger.info("--delete ExerciceController-- exerciceId : " + id);
