+++ src/main/java/fr/treeptik/controller/RoundController.java
@@ -62,7 +62,7 @@ public class RoundController {
 	}
 
 	@RequestMapping(method = RequestMethod.GET, value = "/update/{id}")
-	public String update(Model model, @PathVariable("id") Integer id)
+	public String update(Model model, @PathVariable("id") Long id)
 			throws ControllerException {
 		logger.info("--update RoundController-- roundId : " + id);
 
@@ -79,7 +79,7 @@ public class RoundController {
 	}
 
 	@RequestMapping(method = RequestMethod.GET, value = "/delete/{id}")
-	public String delete(Model model, @PathVariable("id") Integer id)
+	public String delete(Model model, @PathVariable("id") Long id)
 			throws ControllerException {
 		logger.info("--delete RoundController-- roundId : " + id);
