+++ src/main/java/fr/treeptik/dao/ExerciceDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import fr.treeptik.model.Exercice;
 
 public interface ExerciceDAO extends
-		JpaRepository<Exercice, Integer> {
+		JpaRepository<Exercice, Long> {
 
 
 }
