+++ src/main/java/fr/treeptik/dao/NutritionBilanDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import fr.treeptik.model.NutritionBilan;
 
 public interface NutritionBilanDAO extends
-		JpaRepository<NutritionBilan, Integer> {
+		JpaRepository<NutritionBilan, Long> {
 
 
 }
