+++ src/main/java/fr/treeptik/dao/PlatDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import fr.treeptik.model.Plat;
 
 public interface PlatDAO extends
-		JpaRepository<Plat, Integer> {
+		JpaRepository<Plat, Long> {
 
 
 }
