+++ src/main/java/fr/treeptik/service/NutritionBilanService.java
@@ -7,7 +7,7 @@ import fr.treeptik.model.NutritionBilan;
 
 public interface NutritionBilanService {
 
-	NutritionBilan findById(Integer id) throws ServiceException;
+	NutritionBilan findById(Long id) throws ServiceException;
 
 	NutritionBilan create(NutritionBilan nutritionBilan) throws ServiceException;
 
@@ -15,6 +15,6 @@ public interface NutritionBilanService {
 
 	List<NutritionBilan> findAll() throws ServiceException;
 
-	void remove(Integer id) throws ServiceException;
+	void remove(Long id) throws ServiceException;
 
 }
